+++ classes/privacy/provider.php
@@ -18,7 +18,7 @@
  * Privacy subsystem implementation for block_worldclock.
  *
  * @package    block_worldclock
- * @copyright  2026 Adam Jenkins
+ * @copyright  2026 Adam Jenkins <adam@wisecat.net>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */